PlatformSessionFunctions: use int as type for SessionId
It's quite unlikely a user session exists for longer than 68 years.

+++ core/src/PlatformSessionFunctions.h
@@ -32,7 +32,7 @@ class PlatformSessionFunctions
 {
 public:
 	using SessionId = int;
-	using SessionUptime = qint64;
+	using SessionUptime = int;
 
 	struct SessionInfo {
 		SessionId id = InvalidSessionId;
